+++ instrumentor/test/fixtures/is-functional-textinput/output.js
@@ -1,0 +1,6 @@
+var Heap = require('@heap/react-native-heap').default || {
+  withHeapTextInputAutocapture: function withHeapTextInputAutocapture(Component) {
+    return Component;
+  }
+};
+var InternalTextInput = Heap.withHeapTextInputAutocapture(function InternalTextInput(props) {});