+++ instrumentor/test/index.js
@@ -77,6 +77,12 @@ describe('autotrack instrumentor plugin', () => {
     var expected = getExpectedTransformedFile('is-touchable-0-62');
     assert.equal(actual, expected);
   });
+
+  it('functional textinput should be instrumented with HOC', () => {
+    var actual = getActualTransformedFile('is-functional-textinput');
+    var expected = getExpectedTransformedFile('is-functional-textinput');
+    assert.equal(actual, expected);
+  });
 });
 
 const getActualTransformedFile = fixtureName => {